Change current time from uint64_t to double in Date Port API
JerryScript-DCO-1.0-Signed-off-by: Akos Kiss akiss@inf.u-szeged.hu

+++ jerry-core/jerry-port.h
@@ -18,7 +18,6 @@
 #define JERRY_PORT_H
 
 #include <stdbool.h>
-#include <stdint.h>
 #include <stdio.h>
 
 #ifdef __cplusplus
@@ -99,7 +98,7 @@ bool jerry_port_get_time_zone (jerry_time_zone_t *tz_p);
  *
  * @return milliseconds since Unix epoch
  */
-uint64_t jerry_port_get_current_time (void);
+double jerry_port_get_current_time (void);
 
 /**
  * @}
